new post modal

+++ src/components/ViewPost.js
@@ -4,15 +4,18 @@ import { votePost, deletePost } from '../utils/api'
 import * as fn from '../utils/fn'
 import { Container, Icon } from 'semantic-ui-react'
 
-class ViewPost extends Component{
+class ViewPost extends Component {
 
-  state={likes:0}
+  state = { likes: 0, deleted: false }
 
   delete = (postId) => {
-    
+
+    if (!window.confirm('Delete Post?'))
+      return
+
     deletePost(postId).then((result) => {
 
-      console.log("post deleted")
+      this.setState({deleted: result.deleted})
 
     }).catch(error => {
 
@@ -22,13 +25,13 @@ class ViewPost extends Component{
   }
 
   upVote = (postId) => {
-    
-      votePost(postId,"upVote").then((result) => {
 
-        let totalLikes = this.state.likes
-        totalLikes = totalLikes + 1
-        this.setState({likes: totalLikes})
-    
+    votePost(postId, "upVote").then((result) => {
+
+      let totalLikes = this.state.likes
+      totalLikes = totalLikes + 1
+      this.setState({ likes: totalLikes })
+
 
     }).catch(error => {
 
@@ -40,22 +43,22 @@ class ViewPost extends Component{
   }
 
   downVote = (postId) => {
-    
-    votePost(postId,"downVote").then((result) => {
+
+    votePost(postId, "downVote").then((result) => {
 
       let totalLikes = this.state.likes
       totalLikes = totalLikes - 1
-      this.setState({likes: totalLikes})
-  
-
-  }).catch(error => {
-
-    console.log(error);
-
-  })
+      this.setState({ likes: totalLikes })
 
 
-}
+    }).catch(error => {
+
+      console.log(error);
+
+    })
+
+
+  }
 
   render() {
 
@@ -63,30 +66,34 @@ class ViewPost extends Component{
 
     return (
       <Container>
-         
-          {
 
-            <div key={post.id} style={{paddingTop:'1em'}}>
+        {
+          !this.state.deleted && (
+          <div key={post.id} style={{ paddingTop: '1em' }}>
 
+            <div>
+              <h2>
+                <Link to={{ pathname: "/post/" + post.id }}>{post.title}</Link>
+              </h2>
+            </div>
+            <div>
+              <span>Posted by {post.author}</span>
+              <span style={{ paddingLeft: '0.5em' }}>{fn.getDateFormat(post.timestamp)}</span>
+            </div>
+            <div>
               <div>
-                <Link to={{pathname:"/post/" + post.id}}>{post.title}</Link>   
+                {post.voteScore + this.state.likes}<span style={{ paddingLeft: '0.5em' }}>Votes</span>
               </div>
-              <div>
-                {post.voteScore + this.state.likes}
-              </div>
-              <div>
-                {fn.getDateFormat(post.timestamp)}
-              </div>
-              <div>
-                <Icon link name='thumbs up outline' onClick={() => this.upVote(post.id)}></Icon>
-                <Icon link name='thumbs down outline' onClick={() => this.downVote(post.id)}></Icon>
-                <Icon link name='delete' onClick={() => this.delete(post.id)}></Icon>
-                
-              </div>
+
+              <Icon link name='thumbs up outline' onClick={() => this.upVote(post.id)}></Icon>
+              <Icon link name='thumbs down outline' onClick={() => this.downVote(post.id)}></Icon>
+              <Icon link name='delete' onClick={() => this.delete(post.id)}></Icon>
 
             </div>
-            
-          }
+
+          </div>)
+
+        }
 
       </Container>
     )
